+++ test/unit/vagrant/devcommands/command_def_spec.rb
@@ -1,11 +1,34 @@
 require_relative '../../spec_helper'
 
 describe VagrantPlugins::DevCommands::CommandDef do
-  describe 'with missing parameters' do
-    it 'displays raises an ArgumentError' do
-      cmd = described_class.new(name: 'foo', script: 'echo "%s"')
+  describe 'with configured parameters' do
+    it 'interpolates values' do
+      cmd = described_class.new(name:       'foo',
+                                parameters: { that: {} },
+                                script:     'echo "%{that}"')
 
-      expect { cmd.run_script([]) }.to raise_error(ArgumentError)
+      expect(cmd.run_script(['--that', 'works'])).to eq('echo "works"')
+      expect(cmd.run_script(['--that', 'works also'])).to(
+        eq('echo "works also"')
+      )
+    end
+  end
+
+  describe 'with literal percent sign' do
+    it 'needs special escaping' do
+      cmd = described_class.new(name: 'foo', script: 'foo %%s %%bar')
+
+      expect(cmd.run_script([])).to eq('foo %s %bar')
+    end
+  end
+
+  describe 'with missing parameters' do
+    it 'displays raises a KeyError' do
+      cmd = described_class.new(name:       'foo',
+                                parameters: { what: {} },
+                                script:     'echo "%{what}"')
+
+      expect { cmd.run_script([]) }.to raise_error(KeyError)
     end
   end
 end